Adds Router that works with KSP
Router seems to be working right and failing when it should, but it's
yet to be verified if it really works fine with incremental builds.

Tests are failing because for some reason annotated classes are not
being resolved as annotated, specifically in test settings. Tests are
included anyway, and just marked as ignored since the scenarios are
still valid.