Moves publication to JitPack.io
GitHub packages is very user unfriendly for requiring users to pass
credentials to be able to download packages from a public repo. Even if
this wasn't required, each repo has it's own maven repo, which can be
cumbersome anyways. GitHub packages is still interesting for private
artifact sharing, though.

Much of the configuration that was required for GitHub is no longer
needed ans therefore it was removed.

Before I make an official release, I'll make some maintenances on the
router to at least guarantee it works on multimodular projects without
any hacky or suboptimal solution required.