Creates KAPT processor that is capable of multimodule aggregation
KSP can't currently find files/classes from other modules to feed an
aggregating `Router` that needs thoses registrars from dependency
modules in the project. For that reason, I'm begrudgingly implementing
a KAPT port that is capable of doing just that. Was it not for it's
package fetching capabilities, KAPT is much worse to run than KSP felt.

Had to move annotations to the ":crane-annotation" module so I could
use Java `Class<T>` referencing here or there. Offering that module
through `api()` did not seem to do the job, but perhaps it works if
I directly reference a maven uploaded artifact. To be tested. For now
users must explicitly depend on "com.gabrielfv.crane:crane-annotations:"
if they want to use the codegen capabilities.

We will need to revisit tests

Also created an `RouteRegistrar` interface that basicly expose the api
registrars conform to, so we could reference them as
`Set<RouteRegistrar>` instead of `Set<Any>`, which would keep us from
being able to call `registrars.flatMap { it.get() }` since there's no
such `get()` method in `Any` interface.