Further integrates Router with Room Processing
Now we're able to generate files with KotlinPoet through XFiler. This
enables writing kotlin registrars and therefore using the same XStep to
generate registrars for both Kapt and KSP.

However, there are two problems that still keep us from using the same
XStep for the router generation step. The first is that we can't offload
part of the processing back to the actual backend, which would allow for
indirect element fetching in Kapt through `getPackageElement()`. There's
also no equivalent API neither on KSP nor Room Processing so that's a
scenario an XStep is currently incapable of tackling. The other problem
is the current inability of an XStep to defer processing an XElement in
a later round in KSP. This is a behaviour mismatch that would require
some KSP specific extrawork.

Currently to tackle the no-defer problem, we're making the step stateful
and persisted on the KSP processor.