Updates router to more stable API
Makes test work by hacking a way of having a module that just sorta
knows java annotation processing exists and actually knows android
exists. This required creating a new module exclusive for testing and
now tests passes. More tests will be created to validate success
scenarios in the future.

Still there are some quirks related to working with Room Processing and
KSP. It's a matter of both APIs maturing, and I'll be following that.