Creates an API layer on top of the Builer
Now instead of explicitly calling the Builder, it is expected that the
user will be using the CraneRegistry "init" method, staticly exposed via
Crane companion object.

The registry also allows for holding an instance of Crane, which can be
convenient in some DI scenarios. It's still not guaranteed to work if
constructor injection is used in the root fragment because the
initialization process currently creates the first fragment before the
instance is returned. Will see how we can decouple both instances as
transparently as possible in the future.