Moves route wiring to XProcessing
Removes Javac router wiring step and refactors the XProcessing
RoterWiringStep to make use of the new getTypeElementsFromPackage API.
This was one of the feature requests I took part on, made PRs and now am
able to make use of, which makes me happy to see working on my project.

This should enable multi-modular route resolution for KSP as well.